# api/openapi.yaml
openapi: 3.0.3
info:
  title: Kayran Platform API
  description: API для платформы водных активностей, проката, спорта и инклюзивных программ
  version: 0.1.0
  contact:
    name: Kayran Team
servers:
  - url: https://api.kayran.ru/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GPSPoint:
      type: object
      properties:
        lat:
          type: number
          format: float
        lng:
          type: number
          format: float
        timestamp:
          type: string
          format: date-time
        speed:
          type: number
        accuracy:
          type: number
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Регистрация пользователя
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [first_name, last_name, phone]
              properties:
                first_name: { type: string }
                last_name: { type: string }
                phone: { type: string }
                email: { type: string, format: email }
      responses:
        '201':
          description: Пользователь создан
  /auth/login:
    post:
      tags: [Auth]
      summary: Вход в систему
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string }
      responses:
        '200':
          description: JWT-токен

  # Events
  /events:
    get:
      tags: [Events]
      summary: Список событий
      parameters:
        - in: query
          name: event_type
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: point_id
          schema: { type: string }
        - in: query
          name: start_from
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Список событий

    post:
      tags: [Events]
      summary: Создать событие
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [event_type, title, start_time]
              properties:
                event_type: { type: string }
                title: { type: string }
                description: { type: string }
                start_time: { type: string, format: date-time }
                point_id: { type: string }
                route_id: { type: string }
      responses:
        '201':
          description: Событие создано
  /events/{id}:
    get:
      tags: [Events]
      summary: Детали события
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Детали события
  /events/{id}/participants:
    post:
      tags: [Events]
      summary: Добавить участника события
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string }
                role: { type: string }
      responses:
        '200':
          description: Участник добавлен

  # Users
...
  /rentals:
    post:
      tags: [Rentals]
      summary: Создать аренду
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Аренда создана
  /rentals/{id}/return:
    put:
      tags: [Rentals]
      summary: Возврат лодки
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Аренда завершена

  # Routes
  /routes:
    get:
      tags: [Routes]
      summary: Список маршрутов
      parameters:
        - in: query
          name: difficulty
          schema: { type: string }
      responses:
        '200':
          description: Список маршрутов

    post:
      tags: [Routes]
      summary: Создать маршрут
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Маршрут создан
...
  /gps/tracks:
    post:
      tags: [GPS]
      summary: Начать GPS-трекинг
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Трек создан
  /gps/tracks/{id}/points:
    post:
      tags: [GPS]
      summary: Загрузить точки трека
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                points:
                  type: array
                  items:
                    $ref: '#/components/schemas/GPSPoint'
      responses:
        '200':
          description: Точки загружены

  # Sport
...